resolve the commit of the PR

+++ mzLib/FlashLFQ/IsoTracker/XIC.cs
@@ -85,12 +85,12 @@ namespace FlashLFQ.IsoTracker
             var paddedPeaks = new List<IIndexedPeak>();
             var firstPeak = Ms1Peaks[0];
             var lastPeak = Ms1Peaks[Ms1Peaks.Count - 1];
-            double gap = (lastPeak.RetentionTime - firstPeak.RetentionTime) / (Ms1Peaks.Count - 1);
+            float gap = (lastPeak.RetentionTime - firstPeak.RetentionTime) / (Ms1Peaks.Count - 1);
 
             // because we hope to have an odd number of peaks, we have to add the odd number padded peaks
             for (int i = 5; i > 0; i--) //add 4 peaks before the first peak
             {
-                paddedPeaks.Add(new IndexedMassSpectralPeak(0, 0, 0, (float)(firstPeak.RetentionTime - gap * i)));    // not sure about the m/z and index    
+                paddedPeaks.Add(new IndexedMassSpectralPeak(0, 0, 0, firstPeak.RetentionTime - gap * i));    // not sure about the m/z and index    
             }
 
             for (int i = 0; i < Ms1Peaks.Count; i++)
@@ -100,7 +100,7 @@ namespace FlashLFQ.IsoTracker
 
             for (int i = 1; i < 6; i++) //add 5 peaks after the last peak
             {
-                paddedPeaks.Add(new IndexedMassSpectralPeak(0, 0, 0, (float)(lastPeak.RetentionTime + gap * i)));
+                paddedPeaks.Add(new IndexedMassSpectralPeak(0, 0, 0, lastPeak.RetentionTime + gap * i));
             }
 
             return paddedPeaks;
